Add hasura.ts
